Add unavailable conformance to basic types
