+++ Sources/MultipartKit/MultipartFormData.swift
@@ -1,6 +1,11 @@
 import Algorithms
 import Collections
+
+#if canImport(FoundationEssentials)
+import FoundationEssentials
+#else
 import Foundation
+#endif
 
 /// Internal representation of parsed multipart form data with support for hierarchical structures.
 ///
