+++ Tests/MultipartKitTests/Utilities/File.swift
@@ -41,19 +41,9 @@ struct File: Codable, Equatable, MultipartPartConvertible {
     }
 
     init(multipart: MultipartPart<Body>) throws {
-        let contentDisposition = multipart.headerFields[.contentDisposition] ?? ""
-        let filenamePattern = "filename=\"([^\"]+)\""
-        let filename: String
-
-        if let range = contentDisposition.range(of: filenamePattern, options: .regularExpression) {
-            let match = contentDisposition[range]
-            let startIndex = match.index(match.startIndex, offsetBy: 10)  // Skip 'filename="'
-            let endIndex = match.index(before: match.endIndex)  // Skip closing quote
-            filename = String(contentDisposition[startIndex..<endIndex])
-        } else {
+        guard let filename = try multipart.contentDisposition?.filename else {
             throw MultipartError.invalidFileName
         }
-
         self.init(filename: filename, data: multipart.body)
     }
 }
